Modified sms and travel handler
