--- timings.py
@@ -1,150 +1,0 @@
-from collections import OrderedDict
-
-# Legend for personal use only
-bus_legend = dict({"b16ht": "Hamilton - Toronto",
-                   "b16th": "Toronto - Hamilton",
-                   "b15mt": "McMaster - Toronto",
-                   "tLWht": "Hamilton - Toronto",
-                   "tLEth": 'Toronto - Hamilton'
-                   })
-
-b16ht = OrderedDict({'04:15'	:	'05:05'	,
-                     '05:00'	:	'05:55'	,
-                     '05:20'	:	'06:15'	,
-                     '05:40'	:	'06:45'	,
-                     '06:00'	:	'07:15'	,
-                     '06:20'	:	'07:45'	,
-                     '06:40'	:	'08:10'	,
-                     '07:00'	:	'08:35'	,
-                     '07:20'	:	'09:00'	,
-                     '07:40'	:	'09:20'	,
-                     '08:00'	:	'09:35'	,
-                     '08:20'	:	'09:45'	,
-                     '08:40'	:	'10:05'	,
-                     '09:00'	:	'10:20'	,
-                     '09:20'	:	'10:35'	,
-                     '09:40'	:	'10:50'	,
-                     '10:00'	:	'11:05'	,
-                     '10:20'	:	'11:20'	,
-                     '10:40'	:	'11:40'	,
-                     '11:00'	:	'12:00'	,
-                     '11:30'	:	'12:30'	,
-                     '12:00'	:	'13:00'	,
-                     '12:30'	:	'13:30'	,
-                     '13:00'	:	'14:00'	,
-                     '13:30'	:	'14:30'	,
-                     '14:00'	:	'15:00'	,
-                     '14:30'	:	'15:35'	,
-                     '15:00'	:	'16:10'	,
-                     '15:30'	:	'16:45'	,
-                     '16:00'	:	'17:25'	,
-                     '16:30'	:	'18:05'	,
-                     '17:00'	:	'18:35'	,
-                     '17:30'	:	'19:00'	,
-                     '18:00'	:	'19:15'	,
-                     '18:30'	:	'19:40'	,
-                     '19:00'	:	'20:05'	,
-                     '19:30'	:	'20:30'	,
-                     '20:00'	:	'20:55'	,
-                     '20:30'	:	'21:25'	,
-                     '21:00'	:	'21:55'	,
-                     '22:00'	:	'22:55'	,
-                     '23:00'	:	'23:50'	,
-                     '23:30'	:	'00:20'	,
-                     '24:30'	:	'01:20'
-                     })
-
-b16th = OrderedDict({'05:45'	:	'06:40'	,
-                     '06:30'	:	'07:25'	,
-                     '07:00'	:	'08:00'	,
-                     '07:30'	:	'08:55'	,
-                     '08:00'	:	'09:20'	,
-                     '08:30'	:	'09:45'	,
-                     '09:00'	:	'10:05'	,
-                     '09:30'	:	'10:30'	,
-                     '10:00'	:	'11:00'	,
-                     '10:30'	:	'11:30'	,
-                     '11:00'	:	'12:00'	,
-                     '11:30'	:	'12:30'	,
-                     '12:00'	:	'13:00'	,
-                     '12:30'	:	'13:30'	,
-                     '13:00'	:	'14:00'	,
-                     '13:30'	:	'14:35'	,
-                     '14:00'	:	'15:15'	,
-                     '14:30'	:	'15:55'	,
-                     '14:50'	:	'16:25'	,
-                     '15:10'	:	'16:50'	,
-                     '15:30'	:	'17:15'	,
-                     '15:50'	:	'17:35'	,
-                     '16:10'	:	'17:55'	,
-                     '16:30'	:	'18:15'	,
-                     '16:50'	:	'18:35'	,
-                     '17:05'	:	'18:50'	,
-                     '17:20'	:	'19:00'	,
-                     '17:35'	:	'19:10'	,
-                     '17:50'	:	'19:20'	,
-                     '18:05'	:	'19:30'	,
-                     '18:20'	:	'19:40'	,
-                     '18:35'	:	'19:50'	,
-                     '18:50'	:	'20:00'	,
-                     '19:05'	:	'20:15'	,
-                     '19:20'	:	'20:25'	,
-                     '19:40'	:	'20:45'	,
-                     '20:00'	:	'21:00'	,
-                     '20:30'	:	'21:30'	,
-                     '21:00'	:	'22:00'	,
-                     '21:30'	:	'22:25'	,
-                     '22:00'	:	'22:55'	,
-                     '22:30'	:	'23:25'	,
-                     '23:00'	:	'23:55'	,
-                     '23:45'	:	'00:40'	,
-                     '24:30'	:	'01:25'
-                     })
-
-b15mt = OrderedDict({'05:35'	:	'07:00'	,
-                     '06:05'	:	'07:20'	,
-                     '06:33'	:	'07:50'	,
-                     '07:03'	:	'08:20'	,
-                     '07:33'	:	'08:52'	,
-                     '08:06'	:	'09:46'	,
-                     '08:36'	:	'10:16'	,
-                     '09:01'	:	'10:41'	,
-                     '09:38'	:	'11:16'	,
-                     '10:08'	:	'11:46'	,
-                     '10:33'	:	'12:11'	,
-                     '11:03'	:	'12:41'	,
-                     '11:33'	:	'13:11'	,
-                     '12:03'	:	'13:41'	,
-                     '12:33'	:	'14:11'	,
-                     '13:03'	:	'14:41'	,
-                     '13:33'	:	'15:11'	,
-                     '14:03'	:	'15:41'	,
-                     '14:28'	:	'16:08'	,
-                     '14:58'	:	'16:38'	,
-                     '15:31'	:	'17:11'	,
-                     '15:59'	:	'17:35'	,
-                     '16:31'	:	'18:11'	,
-                     '17:01'	:	'18:41'	,
-                     '17:31'	:	'19:11'	,
-                     '18:08'	:	'19:46'	,
-                     '18:33'	:	'20:11'	,
-                     '19:03'	:	'20:41'	,
-                     '19:33'	:	'21:11'	,
-                     '20:03'	:	'21:41'	,
-                     '20:33'	:	'22:11'	,
-                     '21:03'	:	'22:41'	,
-                     '21:33'	:	'23:11'	,
-                     '22:03'	:	'23:41'	,
-                     '22:33'	:	'00:11'	,
-                     '23:03'	:	'00:41'
-                     })
-
-tLWht = OrderedDict({'05:48'	:	'07:00'	,
-                     '06:18'	:	'07:30'	,
-                     '06:48'	:	'08:00'	,
-                     '07:18'	:	'08:30'	})
-
-tLEth = OrderedDict({'16:30'	:	'17:45'	,
-                     '17:00'	:	'18:15'	,
-                     '17:30'	:	'18:45'	,
-                     '18:00'	:	'19:15'	})
